Comments - Defect Number: 3148; Defect RecordID: 3144; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports > Attainment Report Extract > Add new tab for Billable Utilization; Defect Description: We need to add a new third tab to the Attainment report entitled "Billable Utilization". It will become a new source sheet for the Attainment report that is being developed. The extract should display the billable utilization by month, by quarter, and YTD for every Active member of the Logic20/20 consulting organization (excluding internal resources and offshore resources).
We need the following columns in the report. The Month, Quarters, and YTD columns report billable utilization for the individual for that time period in the current year.

Resource Name (Last Name, First Name)
Title
Pay Type
January
February
March
April
May
June
July
August
September
October
November
December
Q1
Q2
Q3
Q4
YTD

Please prioritze building this new extract over all current activities as we would like to be able to use it, if at all possible, for Friday's financial reporting to leadership.

+++ Logic2020, Inc./Practice Management/DotNet/PracticeManagement/ClientDAL/ReportDAL.cs
@@ -2765,7 +2765,7 @@ namespace DataAccess
                     BillableUtlizationByRange billableUtlizationByRange = new BillableUtlizationByRange();
                     billableUtlizationByRange.StartDate = reader.GetDateTime(startDateIndex);
                     billableUtlizationByRange.EndDate = reader.GetDateTime(endDateIndex);
-                    billableUtlizationByRange.BillableUtilization = reader.GetDecimal(billableUtilizationPercentIndex);
+                    billableUtlizationByRange.BillableUtilization = reader.GetDouble(billableUtilizationPercentIndex);
                     billableUtlizationByRange.RangeType = reader.GetString(rangeTypeIndex);
                     attainmentBillableutlizationReport.BillableUtilizationList.Add(billableUtlizationByRange);
                 }
